--- src/main/java/org/icgc_argo/GrpcApplicationService.java
@@ -1,38 +1,0 @@
-/*
- * Copyright (c) 2019. Ontario Institute for Cancer Research
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as
- * published by the Free Software Foundation, either version 3 of the
- * License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <https://www.gnu.org/licenses/>.
- *
- */
-
-package org.icgc_argo;
-
-import org.icgc_argo.util.ProjectBanner;
-import lombok.val;
-import org.springframework.boot.Banner;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.EnableAspectJAutoProxy;
-
-@SpringBootApplication
-@EnableAspectJAutoProxy
-public class GrpcApplicationService {
-
-  public static void main(String[] args) {
-    val application = new SpringApplication(GrpcApplicationService.class);
-    application.setBanner(new ProjectBanner());
-    application.setBannerMode(Banner.Mode.CONSOLE);
-    application.run(args);
-  }
-}
