+++ src/main/java/org/icgc_argo/clinical/grpc/Healthable.java
@@ -1,0 +1,6 @@
+package org.icgc_argo.clinical.grpc;
+
+public interface Healthable {
+
+  String getHealthCheckName();
+}
