--- src/main/java/org/icgc_argo/grpc/Healthable.java
@@ -1,6 +1,0 @@
-package org.icgc_argo.grpc;
-
-public interface Healthable {
-
-  String getHealthCheckName();
-}
