--- src/main/java/org/icgc_argo/model/DriveType.java
@@ -1,33 +1,0 @@
-/*
- * Copyright (c) 2019. Ontario Institute for Cancer Research
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as
- * published by the Free Software Foundation, either version 3 of the
- * License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <https://www.gnu.org/licenses/>.
- *
- */
-
-package org.icgc_argo.model;
-
-import java.util.Optional;
-import java.util.stream.Stream;
-import lombok.NonNull;
-
-public enum DriveType {
-  FRONT,
-  BACK,
-  ALL;
-
-  public static Optional<DriveType> resolveDriveType(@NonNull String typeName) {
-    return Stream.of(values()).filter(x -> x.name().equals(typeName)).findFirst();
-  }
-}
